Debugging on orin; cstring error in unified_can_driver.cpp

+++ ros2_ws/src/unified_can_driver/src/module/dvl.cpp
@@ -3,7 +3,7 @@
 using namespace Module;
 
 DVLModule::DVLModule(rclcpp::Node* ctx, Mailbox::MboxCan* mb)
-	: DeviceModule(ctx, mb, 10, CanDriver::Device::WAYFDVL::ID, true, true, 16)
+	: DeviceModule(ctx, mb, 10, static_cast<uint8_t>(CanDriver::Device::WAYFDVL::ID), true, true, 16)
 {
 	/* pointer-to-member-function weirdness dictates that each topic function is manually added*/
 	module_topic_ptr_array[0x0] = static_cast<topic_ptr_t>(&DVLModule::dres_info);			/* 0000 NOP */
@@ -38,7 +38,7 @@ DVLModule::DVLModule(rclcpp::Node* ctx, Mailbox::MboxCan* mb)
 	struct can_frame init_frame;
 	memset(&init_frame, 0, sizeof(struct can_frame));
 	init_frame.can_dlc = 4;
-	init_frame.can_id = CanDriver::Command::STOW;
+	init_frame.can_id = static_cast<uint8_t>(CanDriver::Command::STOW);
 	init_frame.data[0] = this->module_device_id;
 	Mailbox::MboxCan::write_mbox(this->mailbox_ptr, &init_frame);
 
@@ -69,10 +69,10 @@ void DVLModule::timer_callback()
 	{
 		struct can_frame poll_frame;
 		memset(&poll_frame, 0, sizeof(struct can_frame));
-		poll_frame.can_id = CanDriver::Command::DREQ;
+		poll_frame.can_id = static_cast<uint8_t>(CanDriver::Command::DREQ);
 		poll_frame.can_dlc = 4;
 		poll_frame.data[0] = this->module_device_id;
-		poll_frame.data[2] = CanDriver::Device::WAYFDVL::TOPIC_VEL_AND_DIST;
+		poll_frame.data[2] = static_cast<uint8_t>(CanDriver::Device::WAYFDVL::TOPIC_VEL_AND_DIST);
 		Mailbox::MboxCan::write_mbox(mailbox_ptr, &poll_frame);
 	}
 }
